Randomseed as of the morning of 1 May 2023
Randomseed generator is working.
The code is slow, and took 72 hours to train an agent on 100 epochs of randomseed board.  I have started to optimize the code (by including the wire extension as a method within the randomseed generator), but there is a lot more optimization to be done.
The saved randomseed-trained agent performs middling well when tested against all the various boards, better than the offline-BFSshort or offline-L-systems agents, but worse than the offline-randomwalk or offline-numberlink or the uniform-trained agents.